app/requirements.txt
Django==5.1.2
gunicorn==21.2.0
psycopg2-binary
djangorestframework==3.15.2
drf-writable-nested==0.7.1
djangorestframework-simplejwt==5.3.1
dask==2024.5.0
dask-expr==1.1.0
pyperf==2.8.0
dj-rest-auth==7.0.0
django-filter==24.3
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
django-cors-headers==4.6.0
django-safedelete==1.4.0
celery[redis]==5.4.0
django-celery-results==2.5.1
flower==2.0.1
watchfiles==1.0.4
djoser==2.3.1
redis==5.2.1
django-debug-toolbar==5.0.1
whitenoise==6.9.0
pandas
drf-proxy-pagination
drf-access-policy===1.5.0
